First commit for approach 1:
The idea behind this approach is to load all the data from the file to
the database, and after that, query the database for the request between
the given time range (startDate + duration). So, after this, I will
count those request that meet the threshold and I will saved to database
in a different table and printed them to the console as well.